Add basic scaffolding and Range class
This change introduces the basic scaffolding of the package along with
the first class, Range. A range in our sense is a tuple of two years, a
'first' year marking the beginning of a time span, and a 'last' year
representing the end of a time span. A Range imposes a single constraint
over the two years in that the first year must be less or equal to the
last year.
Along with this initial first class the change also adds a test case for
the enforcement of this aforementioned constraint.