Add string conversion support for Range objects
A range is a more or less abstract concept. As it stands, however, there
is no way to represent such a range outside of a Python program.
This change introduces a string conversion function allowing a Range
object to be converted into a string (which later can be written into a
file) as we expect it to be used for a span of a start and an end year,
separated by a dash. E.g., the span of the years 2000 to 2015 would be
represented as the string '2000-2015'.