Add support for checking inclusion of years
A Range representing a time span (with the granularity being full
years), it seems natural to be able to check if another year is included
in any given range.
This patch adds support for such a check. It introduces a __contains__
method which allows for idiomatic Python code using the 'in' operator
allowing for inclusion checks on other types such a lists.