Add functionality to check if a year can extend a range
In the future we want to merge ranges of years in order to minimize the
size of a copyright string. To that end, we need to be able to check if
a given year can extend a range, i.e., whether it could replace one of
the Range's years without introducing a semantical change (by increasing
the effective time span represented).
This change introduces a new method, extendedBy, to the Range class that
can be used whether a given year extends a range in the aforementioned
sense. Note that we only regard extension towards the upper end of the
time span represented by the year, i.e., the 'last' year of the tuple.
This constraint is supposed to simplify handling later on.