Introduce range parsing support
As it stands we have basic support for working with Range objects but we
have no real way of creating them based on some input data. In the
future we anticipate to get passed in a file and to have to parse a
string representing the years of a copyright into a set of ranges. To
work towards that goal, this change adds support for parsing a single
range, i.e., creating a Range object out of a string. A range can be
represented in two different ways: If it represents a single year it
will be encoded as a single number which can directly be parsed as an
integer. A range spanning multiple years is represented as two numbers
by a dash and is parsed as two individual numbers there were retrieved
by splitting the initial string at the dash.