Add Range comparison support
An upcoming change will introduce support for normalizing a list of
ranges by merging them. The process of detecting whether or not two
ranges can be merged is greatly simplified if the ranges are ordered
according to some means.
This change introduces functionality for comparing two ranges which
directly leads to the ability to sort a list of Range objects, and
hence, carving the way for the mentioned future work.