Add support for normalizing a list of Range objects
For sanitizing the copyright year string we want to be able to handle
wrongly specified year specifications (to a degree where we can still
tell the intention) and also have the ability to extend the string by
the current year. In our scenario, the string would first be parsed into
a set (actually: list) of Range objects representing the various years
where a copyright was claimed.
This change adds an algorithm that can be used for normalizing a list of
ranges. Normalization in this sense means the creation of the smallest
set of ranges that still preserves the semantics of the original. The
algorithm works as follows: First we sort the range list. Now we iterate
over the ranges and merge every two that overlap into a single one.
Because the list was sorted, a single pass is enough to fully normalize
the list.