Add range sequence parsing support
In the end, we will be given a string representing the years of
copyright. We need to be able to represent these years as Range objects
in Python.
To achieve this goal, this patch adds support for parsing such a string
into a sequence of ranges. The implementation is straightforward in that
we mainly invoke the Range's parse() method on the individual range
strings after some pre-processing and splitting of the original
copyright year string.