Add normalize.py script
Until now we created infrastructure used for supporting our anticipated
use-case but aside from tests nothing made use of the created
functionality.
This change finally adds a script that is able to parse, normalize, and
update the copyright years in a file. The script accepts a list of
files. It will read each of them on a line-by-line basis. Each line, in
turn, is scanned for a pattern containing the string "Copyright"
(because that is the most commonly used indicator of a copyright
header). After this string, a string containing the copyright years is
expected (we use some sort of fuzzy matching to retrieve that). This
string is subsequently parsed into a list of regions. The current year
is added to the list, the list is normalized, and the new copyright line
written into the file.