Add basic policy support
For the most part, normalization can be seen as a policy (as opposed to
a mechanism): There is no true wrong or right policy and while some
policies are more flexible in what they can do, none will be able to
cover all cases and fulfill all requirements. To that end, we want to be
able to work with different policies, which may differ in the input
understood or the output produced.
This change introduces a new parameter, --policy, to the normalization
infrastructure. Currently we only have a single policy, 'plain' (the
default one), but there are more to come.