Introduce 'pad' policy
Until now the normalization was bound to a single algorithm for
normalizing copyright years. Arguably, there could potentially be many
different formats that should be understood and similarly many formats
that we could normalize to. While we only care for one output format,
some copyright headers are special and our replacement algorithm
produces output that cannot be considered correct. In particular,
headers that are framed will not be handled properly once we change the
length of a copyright year string.
This change introduces a new policy that aims at correctly handling
framed copyright headers. In essence, it performs the normalization of
copyright years as before but additionally it tries to pad the string
next to the copyright years (on the right) by as many spaces as were
removed previously or to remove as many spaces as characters were added,
respectively.