Add support for requiring existence of a copyright header
Not all source code files might contain copyright headers. Some of them
accidentally so. We might want to detect those cases.
This change causes the copyright header substitution to return the
number of replacements made. This information can be used to infer the
number of copyright headers in a file.