Allow passing in of year to extend copyright range with
Previously, the normalization functions implicitly extended the
copyright year range parsed from any file content with the current year.
This approach is not very flexible. Rather, the module should allow for
specifying the year to extend with, if any.
This change adjusts the module's normalization functions to accept a
'year' parameter specifying the year to extend a copyright range with.
Because of this change the testing is simplified as well, alleviating
the need for the mock module employed beforehand: whereas previously we
had to patch the datetime query at runtime in order to extend the
copyright string with a fixed year we know simply pass in the year of
interest.