Add copyright header ignore support
When working with code (partly) contributed by others an additional
copyright headers is usually present identifying the original copyright
holders. Those additional headers should not be updated, and likely also
not normalized.
To handle those cases this change introduces the infrastructure to
ignore headers matching a certain pattern (i.e., by maintaining a
blacklist). It introduces a new option, --ignore, that accepts a pattern
the matched copyright header is matched against. If a match was found
no action will be performed. The option can be specified multiple times.